chore: add paper and fix styles
