+++ src/components/atoms/Paper/Paper.component.tsx
@@ -1,0 +1,18 @@
+import React from 'react'
+import Paper from './Paper.style'
+
+type PaperComponentInterface = {
+  children?: JSX.Element | JSX.Element[]
+};
+
+const PaperComponent: React.FC<
+  PaperComponentInterface
+> = ({
+  children,
+}) => (
+  <Paper>
+    {children}
+  </Paper>
+)
+
+export default PaperComponent
