+++ src/components/atoms/Paper/Paper.style.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+const Paper = styled.div`
+  background-color: black;
+  color: white;
+`
+
+export default Paper
