+++ src/components/atoms/Paper/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Paper.component'
