+++ src/pages/Home/index.tsx
@@ -1,17 +1,1 @@
-import React from 'react'
-
-import Container from '../../components/atoms/Container'
-import Button from '../../components/atoms/Button'
-
-const Home = () => (
-  <Container>
-    <>
-      <h1>Home</h1>
-      <Button>
-        Login
-      </Button>
-    </>
-  </Container>
-)
-
-export default Home
+export { default } from './Home.component'
